Move RollbackToSnapshotProcedure package
